4-Session/requirement.txt
pymupdf
requests
python-dotenv
openai